+++ ca/django_ca/managers.py
@@ -243,7 +243,9 @@ class CertificateAuthorityManager(
             <settings-ca-default-expires>`.
         algorithm : :py:class:`~cg:cryptography.hazmat.primitives.hashes.HashAlgorithm`, optional
             Hash algorithm used when signing the certificate, defaults to
-            :ref:`CA_DIGEST_ALGORITHM <settings-ca-digest-algorithm>`.
+            :ref:`settings-ca-default-signature-hash-algorithm` for RSA/EC keys, and
+            :ref:`settings-ca-dsa-signature-hash-algorithm` for DSA keys. Passing an algorithm for
+            Ed448/Ed25519 keys is an error.
         parent : :py:class:`~django_ca.models.CertificateAuthority`, optional
             Parent certificate authority for the new CA. Passing this value makes the CA an intermediate
             authority. Let unset if this CA will be used for OpenSSH.